Split test units down to individual files
